[web] Move frame_reference.dart to the html/ folder
