+++ engine/src/flutter/lib/web_ui/lib/src/engine.dart
@@ -65,7 +65,6 @@ export 'engine/font_change_util.dart';
 export 'engine/font_fallback_data.dart';
 export 'engine/font_fallbacks.dart';
 export 'engine/fonts.dart';
-export 'engine/frame_reference.dart';
 export 'engine/frame_timing_recorder.dart';
 export 'engine/html/backdrop_filter.dart';
 export 'engine/html/bitmap_canvas.dart';
@@ -74,6 +73,7 @@ export 'engine/html/clip.dart';
 export 'engine/html/color_filter.dart';
 export 'engine/html/debug_canvas_reuse_overlay.dart';
 export 'engine/html/dom_canvas.dart';
+export 'engine/html/frame_reference.dart';
 export 'engine/html/image.dart';
 export 'engine/html/image_filter.dart';
 export 'engine/html/offset.dart';
