+++ engine/src/flutter/lib/web_ui/lib/src/engine/html/bitmap_canvas.dart
@@ -12,13 +12,13 @@ import '../canvas_pool.dart';
 import '../display.dart';
 import '../dom.dart';
 import '../engine_canvas.dart';
-import '../frame_reference.dart';
 import '../text/canvas_paragraph.dart';
 import '../util.dart';
 import '../vector_math.dart';
 import 'clip.dart';
 import 'color_filter.dart';
 import 'dom_canvas.dart';
+import 'frame_reference.dart';
 import 'image.dart';
 import 'painting.dart';
 import 'path/path.dart';
