+++ engine/src/flutter/lib/web_ui/lib/src/engine/html/picture.dart
@@ -10,12 +10,12 @@ import 'package:ui/ui.dart' as ui;
 
 import '../dom.dart';
 import '../engine_canvas.dart';
-import '../frame_reference.dart';
 import '../util.dart';
 import '../vector_math.dart';
 import 'bitmap_canvas.dart';
 import 'debug_canvas_reuse_overlay.dart';
 import 'dom_canvas.dart';
+import 'frame_reference.dart';
 import 'image.dart';
 import 'path/path_metrics.dart';
 import 'recording_canvas.dart';
